Changing Item Entity Provider to return tem.getUnlocalizedName for unlocalizedName.
This is to ensure we're matching the correct filtername for the item.  Before this change, for Wither Skulls, the filter would only filter on "item.skull.wither" but the unlocalizedname variable, type and name varaible were all something different.  Not sure why we originally were doing tem.getItem() as that returns the item.skull instead of item.skull for the unlocalizedname.